requirements.txt
async-timeout==4.0.3
chardet==5.2.0
clarifai==10.5.3
configparser==7.0.0
curlify==2.2.1
EasyProcess==1.1
email_validator==2.2.0
emoji==2.12.1
fastapi-cli==0.0.4
fqdn==1.5.1
future==1.0.0
google-analytics==0.0.0
google-analytics-data==0.18.9
google-cloud-bigquery==3.25.0
googleAPI==0.1.0
grpcio-status==1.65.1
httptools==0.6.1
huepy==1.2.1
isoduration==20.11.0
jsonpointer==3.0.0
jupyter==1.0.0
jupyter_server_proxy==4.3.0
jupyterlab_vpython==3.1.8
jwt==1.3.1
MeaningCloud-python==2.0.0
mock==5.1.0
orjson==3.10.5
passlib==1.7.4
plyer==2.1.0
pycountry==24.6.1
pycryptodome==3.20.0
PyJWT==2.8.0
pymongo==4.7.3
pyspring==0.0.2
pytest==8.2.2
python-dotenv==1.0.1
python-jose==3.3.0
PyVirtualDisplay==3.0
regex==2024.5.15
responses==0.25.3
schedule==1.2.2
selenium==4.22.0
supabase==2.5.0
ujson==5.10.0
uri-template==1.3.0
uvloop==0.19.0
webcolors==24.6.0
PyYAML==6.0.1
reflex==0.5.10
reflex-chakra==0.5.10
serpapi==0.1.5